Move logic to PaywallCoordinator+Model
Testable later.
